requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
azure-cognitiveservices-speech==1.40.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.1
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.0
cors==1.0.1
cvzone==1.6.1
cycler==0.12.1
Cython==3.1.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2025.2.0
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
Django==5.1
docker==7.1.0
docx2txt==0.9
et_xmlfile==2.0.0
faiss-cpu==1.11.0
fastapi==0.115.5
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fpdf==1.7.2
frozenlist==1.6.0
fsspec==2025.2.0
future==1.0.0
gast==0.6.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
globe==1.0.52
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
gunicorn==23.0.0
h11==0.16.0
h5py==3.12.1
holidays==0.74
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.33
jaxlib==0.4.33
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonify==0.5
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.10.0
kiwisolver==1.4.7
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
lime==0.2.0.1
llvmlite==0.43.0
locket==1.0.0
lxml==5.4.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
mdurl==0.1.2
media==0.1.6
mediapipe==0.10.14
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.26.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
openai==1.79.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.12.1
orjson==3.10.18
packaging==24.1
pandas==2.2.3
partd==1.4.2
patsy==1.0.1
pillow==10.4.0
pipe==2.2
platformdirs==4.3.6
plotly==6.0.0
pmdarima==2.0.4
pooch==1.8.2
propcache==0.3.1
prophet==1.1.7
proto-plus==1.26.0
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycurl==7.45.4
pydantic==2.10.0
pydantic-settings==2.9.1
pydantic_core==2.27.0
pydeck==0.9.1
pyFirmata==1.1.0
Pygments==2.18.0
pyparsing==3.1.4
pypdf==5.5.0
PyPDF2==3.0.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2024.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.0
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentence-transformers==4.1.0
setuptools==75.1.0
shap==0.47.0
six==1.16.0
slicer==0.0.8
smmap==5.0.2
smote==0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.13.0
soxr==0.5.0.post1
SQLAlchemy==2.0.41
sqlparse==0.5.1
stanio==0.5.1
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.45.1
sympy==1.14.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.7.0
torchvision==0.22.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.135
ultralytics-thop==2.0.14
urllib3==2.2.3
uvicorn==0.32.1
watchdog==6.0.0
Werkzeug==3.0.6
wheel==0.44.0
wrapt==1.16.0
xgboost==2.1.3
XlsxWriter==3.2.2
yarl==1.20.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
